+++ Assets/Scripts/UI/UIHighlightClusterHandler.cs
@@ -1,7 +1,8 @@
+using System.Collections.Generic;
 using Cysharp.Threading.Tasks;
-using Unity.Collections;
 using Unity.Entities;
 using UnityEngine;
+using InputCamera;
 using Enemy.ECS;
 using Gameplay;
 
@@ -9,22 +10,38 @@ namespace UI
 {
     public class UIHighlightClusterHandler : MonoBehaviour 
     {
-        // Needs to handle unhovering and re-hovering current highlighted cluster
+
+        [SerializeField]
+        private UIHighlightClusterDisplay displayPrefab;
+
+        [SerializeField]
+        private RectTransform displayParent;
+        
+        private List<UIHighlightClusterDisplay> spawnedDisplays = new List<UIHighlightClusterDisplay>();
         
         private GameManager gameManager;
+        private InputManager inputManager;
         
-        private EntityManager entityManager;
+        private HighlightClusterDataComponent? lastData;
         private EntityQuery highlightClusterQuery;
-
-        private bool reading;
+        private EntityManager entityManager;
+        
+        private bool displaying;
         
         private void Awake()
         {
             entityManager = World.DefaultGameObjectInjectionWorld.EntityManager;
             highlightClusterQuery = entityManager.CreateEntityQuery(typeof(HighlightClusterDataComponent));
+            
             GetGameManager().Forget();
+            GetInput().Forget();
         }
 
+        private async UniTaskVoid GetInput()
+        {
+            inputManager = await InputManager.Get();
+        }
+        
         private async UniTaskVoid GetGameManager()
         {
             gameManager = await GameManager.Get();
@@ -32,48 +49,79 @@ namespace UI
 
         private void Update()
         {
-            if (reading || gameManager.IsGameOver)
+            if (inputManager == null || gameManager.IsGameOver || InputManager.MouseOverUI())
             {
                 return;
             }
             
-            ReadHighlightClusterData().Forget();
+            ReadHighlightClusterData();
+
+            HandleDisplaying();
         }
         
-        private async UniTaskVoid ReadHighlightClusterData()
+        private void ReadHighlightClusterData()
         {
-            reading = true;
-            
-            NativeList<HighlightClusterDataComponent> array = highlightClusterQuery.ToComponentDataListAsync<HighlightClusterDataComponent>(Allocator.TempJob, out var awaitJobHandle);
-            awaitJobHandle.Complete();
-            while (!awaitJobHandle.IsCompleted)
-            {
-                await UniTask.Yield();
-            }
-            reading = false;
-
-            if (!array.IsCreated) return;
-
-            switch (array.Length)
+            int entityCount = highlightClusterQuery.CalculateEntityCount();
+            switch (entityCount)
             {
                 case 0:
-                    array.Dispose();
                     return;
                 case > 1:
-                    Debug.LogError("Should not exceed length 1!!!");
                     entityManager.DestroyEntity(highlightClusterQuery);
-                    array.Dispose();
                     return;
             }
 
-            foreach (HighlightClusterDataComponent cluster in array)
+            if (!highlightClusterQuery.TryGetSingleton(out HighlightClusterDataComponent highlightData)) return;
+            
+            lastData = highlightData;
+            if (displaying)
             {
-                
+                HideDisplays();
+                Display(highlightData);
             }
             
             entityManager.DestroyEntity(highlightClusterQuery);
+        }
+        
+        private void HandleDisplaying()
+        {
+            if (!lastData.HasValue)
+            {
+                return;
+            }
+
+            if (inputManager.CurrentMousePathIndex.Equals(lastData.Value.PathIndex))
+            {
+                if (!displaying)
+                {
+                    Display(lastData.Value);
+                }
+            }
+            else if (displaying)
+            {
+                HideDisplays();
+            }
+        }
+        
+        private void Display(HighlightClusterDataComponent highlightData)
+        {
+            displaying = true;
             
-            array.Dispose();
+            UIHighlightClusterDisplay display = displayPrefab.Get<UIHighlightClusterDisplay>(displayParent);
+            display.Display(highlightData);
+            spawnedDisplays.Add(display);
+        }
+
+        private void HideDisplays()
+        {
+            displaying = false;
+
+            for (int i = 0; i < spawnedDisplays.Count; i++)
+            {
+                spawnedDisplays[i].Hide();
+            }
+            
+            spawnedDisplays.Clear();
         }
     }
 }
